Corrected type

+++ FrameLib_Framework/FrameLib_Threading.cpp
@@ -91,7 +91,7 @@ void Semaphore::close()
         
         // Signal until the count is zero (only reliable way to signal all waiting threads
         
-        long releaseCount = 0;
+        int releaseCount = 0;
         sem_getvalue(&mInternal, &releaseCount);
         
         for (; releaseCount; --releaseCount)
